requirements.txt
python-dotenv
openai
langfuse